transform hashes now per entry, covering all scenarios now.

+++ include/vierkant/nodes.hpp
@@ -60,17 +60,17 @@ NodeConstPtr node_by_name(NodeConstPtr root, const std::string &name);
  */
 void build_node_matrices(const NodeConstPtr &root,
                          const node_animation_t &animation,
-                         float animation_time,
+                         double animation_time,
                          std::vector<glm::mat4> &matrices);
 
 void build_node_matrices_bfs(const NodeConstPtr &root,
                              const node_animation_t &animation,
-                             float animation_time,
+                             double animation_time,
                              std::vector<glm::mat4> &matrices);
 
 void build_morph_weights_bfs(const NodeConstPtr &root,
                              const node_animation_t &animation,
-                             float animation_time,
+                             double animation_time,
                              std::vector<std::vector<float>> &morph_weights);
 
 }// namespace vierkant::bones
